Gateway/requirements.txt
adafruit-io==2.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
paho-mqtt==1.6.1
pyserial==3.5
requests==2.27.1
urllib3==1.26.9
